refactor: use new CidHashSet implementation
