+++ src/ipld/mod.rs
@@ -1,7 +1,6 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-mod cid_hashset;
 mod frozen_cids;
 pub mod json;
 pub mod selector;
@@ -11,7 +10,6 @@ pub use libipld::Path;
 pub use libipld_core::ipld::Ipld;
 pub use util::*;
 
-pub use self::cid_hashset::CidHashSet;
 pub use self::frozen_cids::FrozenCids;
 
 pub use libipld_core::serde::{from_ipld, to_ipld};
