+++ src/tool/subcommands/archive_cmd.rs
@@ -6,10 +6,11 @@ use crate::chain::{
     index::{ChainIndex, ResolveNullTipset},
     ChainEpochDelta,
 };
+use crate::cid_collections::CidHashSet;
 use crate::cli_shared::{snapshot, snapshot::TrustedVendor};
 use crate::db::car::ManyCar;
 use crate::db::car::{AnyCar, RandomAccessFileReader};
-use crate::ipld::{stream_graph, unordered_stream_graph, CidHashSet};
+use crate::ipld::{stream_graph, unordered_stream_graph};
 use crate::networks::{calibnet, mainnet, ChainConfig, NetworkChain};
 use crate::shim::clock::{ChainEpoch, EPOCHS_IN_DAY, EPOCH_DURATION_SECONDS};
 use anyhow::{bail, Context as _, Result};
